Display all time entries

+++ lib/modules/task_selection/ui/time_entries_list/time_entries_list_bloc.dart
@@ -1,10 +1,12 @@
 import 'package:flutter/material.dart';
 import 'package:freezed_annotation/freezed_annotation.dart';
+import 'package:intl/intl.dart';
 import 'package:open_project_time_tracker/app/auth/domain/auth_service.dart';
 import 'package:open_project_time_tracker/app/ui/bloc/bloc.dart';
 import 'package:open_project_time_tracker/modules/task_selection/domain/settings_repository.dart';
 import 'package:open_project_time_tracker/modules/task_selection/domain/time_entries_repository.dart';
 import 'package:open_project_time_tracker/modules/timer/domain/timer_repository.dart';
+import 'package:collection/collection.dart';
 
 import '../../../calendar/domain/calendar_notifications_service.dart';
 
@@ -14,7 +16,7 @@ part 'time_entries_list_bloc.freezed.dart';
 class TimeEntriesListState with _$TimeEntriesListState {
   const factory TimeEntriesListState.loading() = _Loading;
   const factory TimeEntriesListState.idle({
-    required List<TimeEntry> timeEntries,
+    required Map<DateTime, List<TimeEntry>> timeEntries,
     required Duration workingHours,
     required Duration totalDuration,
   }) = _Idle;
@@ -34,16 +36,6 @@ class TimeEntriesListBloc
   final AuthService _graphAuthService;
   final TimerRepository _timerRepository;
   final CalendarNotificationsService _calendarNotificationsService;
-
-  List<TimeEntry> items = [];
-  Duration workingHours = const Duration(hours: 0);
-  Duration get totalDuration {
-    var result = const Duration();
-    for (var element in items) {
-      result += element.hours;
-    }
-    return result;
-  }
 
   TimeEntriesListBloc(
     this._timeEntriesRepository,
@@ -70,42 +62,49 @@ class TimeEntriesListBloc
     super.didChangeAppLifecycleState(state);
   }
 
-  Future<void> reload({
-    bool showLoading = false,
-  }) async {
+  Future<void> reload({bool showLoading = false}) async {
     try {
       if (showLoading) {
         emit(const TimeEntriesListState.loading());
       }
-      items = await _timeEntriesRepository.list(
-        userId: 'me',
-        startDate: DateTime.now(),
-        endDate: DateTime.now(),
+      final items = await _timeEntriesRepository.list(userId: 'me');
+      final mappedItems = items.groupListsBy(
+        (elements) =>
+            DateTime.parse(DateFormat('yyyy-MM-dd').format(elements.spentOn)),
       );
-      workingHours = await _settingsRepository.workingHours;
-      emit(TimeEntriesListState.idle(
-        workingHours: workingHours,
-        timeEntries: items,
-        totalDuration: totalDuration,
-      ));
+      final workingHours = await _settingsRepository.workingHours;
+      emit(
+        TimeEntriesListState.idle(
+          workingHours: workingHours,
+          timeEntries: mappedItems,
+          totalDuration: _totalDuration(mappedItems),
+        ),
+      );
     } catch (e) {
-      emit(TimeEntriesListState.idle(
-        workingHours: workingHours,
-        timeEntries: [],
-        totalDuration: const Duration(),
-      ));
+      emit(
+        TimeEntriesListState.idle(
+          timeEntries: {},
+          workingHours: const Duration(),
+          totalDuration: const Duration(),
+        ),
+      );
       emitEffect(const TimeEntriesListEffect.error());
     }
   }
 
   Future<void> updateWorkingHours(Duration value) async {
     await _settingsRepository.setWorkingHours(value);
-    workingHours = value;
-    emit(TimeEntriesListState.idle(
-      workingHours: workingHours,
-      timeEntries: items,
-      totalDuration: totalDuration,
-    ));
+    state.whenOrNull(
+      idle: (timeEntries, workingHours, totalDuration) {
+        emit(
+          TimeEntriesListState.idle(
+            timeEntries: timeEntries,
+            workingHours: value,
+            totalDuration: totalDuration,
+          ),
+        );
+      },
+    );
   }
 
   Future<void> unauthorize() async {
@@ -116,30 +115,43 @@ class TimeEntriesListBloc
     await _authService.logout();
   }
 
-  Future<void> setTimeEntry(
-    TimeEntry timeEntry,
-  ) async {
-    await _timerRepository.setTimeEntry(
-      timeEntry: timeEntry,
-    );
+  Future<void> setTimeEntry(TimeEntry timeEntry) async {
+    await _timerRepository.setTimeEntry(timeEntry: timeEntry);
   }
 
   Future<bool> deleteTimeEntry(int id) async {
-    try {
-      await _timeEntriesRepository.delete(id: id);
-      items.removeWhere((element) => element.id == id);
-
-      Future.delayed(const Duration(milliseconds: 250)).then((value) {
-        emit(TimeEntriesListState.idle(
-          workingHours: workingHours,
-          timeEntries: items,
-          totalDuration: totalDuration,
-        ));
-      });
-      return true;
-    } catch (e) {
-      emitEffect(const TimeEntriesListEffect.error());
+    if (state is _Idle) {
+      final idleState = state as _Idle;
+      try {
+        await _timeEntriesRepository.delete(id: id);
+        idleState.timeEntries.forEach((key, value) {
+          value.removeWhere((element) => element.id == id);
+        });
+        Future.delayed(const Duration(milliseconds: 250)).then((value) {
+          emit(
+            TimeEntriesListState.idle(
+              workingHours: idleState.workingHours,
+              timeEntries: idleState.timeEntries,
+              totalDuration: idleState.totalDuration,
+            ),
+          );
+        });
+        return true;
+      } catch (e) {
+        emitEffect(const TimeEntriesListEffect.error());
+        return false;
+      }
+    } else {
       return false;
     }
+  }
+
+  Duration _totalDuration(Map<DateTime, List<TimeEntry>> items) {
+    final todaysItems = items.entries.firstOrNull?.value ?? [];
+    var result = const Duration();
+    for (var element in todaysItems) {
+      result += element.hours;
+    }
+    return result;
   }
 }
